+++ lapce-proxy/src/buffer.rs
@@ -54,7 +54,11 @@ impl Buffer {
                 ext
             },
         );
-        let path = self.path.canonicalize()?;
+        let path = if self.path.is_symlink() {
+            self.path.canonicalize()?
+        } else {
+            self.path.clone()
+        };
         let tmp_path = &path.with_extension(tmp_extension);
 
         let mut f = File::create(tmp_path)?;
@@ -235,7 +239,7 @@ pub fn language_id_from_path(path: &Path) -> Option<&'static str> {
                     "tsx" => "typescriptreact",
                     "tex" => "tex",
                     "vb" => "vb",
-                    "xml" => "xml",
+                    "xml" | "csproj" => "xml",
                     "xsl" => "xsl",
                     "yml" | "yaml" => "yaml",
                     "zig" => "zig",