+++ lapce-ui/src/explorer.rs
@@ -1267,7 +1267,7 @@ impl OpenEditorList {
             }
             EditorTabChild::Settings { .. } => {
                 text = "Settings".to_string();
-                hint = format!("ver. {}", *meta::VERSION);
+                hint = format!("ver. {}", meta::VERSION);
                 svg = data.config.ui_svg(LapceIcons::SETTINGS);
             }
             EditorTabChild::Plugin { volt_name, .. } => {